modify structure
