+++ src/class02_03/Code03_DoubleEndsQueueToStackAndQueue.java
@@ -1,4 +1,4 @@
-package class02;
+package class02_03;
 
 import java.util.LinkedList;
 import java.util.Queue;
