+++ src/class02_03/Code06_TwoStacksImplementQueue.java
@@ -1,4 +1,4 @@
-package class02;
+package class02_03;
 
 import java.util.Stack;
