+++ src/class02_03/Code08_GetMax.java
@@ -1,4 +1,4 @@
-package class02;
+package class02_03;
 
 public class Code08_GetMax {
