+++ src/class02_03/HashMapAndSortedMap.java
@@ -1,4 +1,4 @@
-package class02;
+package class02_03;
 
 import java.util.HashMap;
 import java.util.HashSet;
